2023 Day 20
Kinda a pain in the ass. I finished on the night of but wanted to
clean up my code a bit before committing it, but in doing so I broke
something. So I had to go back and fix it today. Also did part 1
of 21 but part 2 is deep fried bullshit and I want to get my solution
on it working even though I did read about the math trick and it would
be easier just doing that.