2023 FINISHED! and some clean up
First time finishing aoc. Got close in past but holidays ramping up
always got in the way at some point. Did days 23 and 24 on the days
of, only just now got around to fixing my day 25. Got ideas and help
from the community for the last two.

* Day 23 I ended up just brute forcing part 2 with my part 1 accidentally.
I ran it on input and after a bit of time I started thinking about a plan
but it ended up finishing a few minutes later so I just used the answer.

* Day 24 was annoying mostly because seeing that people were using z3 got
me to look into z3 packages for go. Ended up struggling for a while trying to
get it to work even though the packages are old and I've never used z3 before.
Gave up and just plugged in equations to an online solver.

* Day 25 was also a headache. Didn't get it finished til now. Was doing
Karger's but for some reason it wasn't working on the input. Then I tried
a few different approaches all without luck. Then life got in way as always
Ultimately found out the reason it failed was I was consolidating edges
that I shouldn't have been.